+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v49.x / v50 (In progress)
+- PC sprites: Added Exit Mouse, Duster, Save Frog, Edgar, Rek, Dorothy, and Sophia.
+
 ## v49 (2021-10-11)
 - PC sprites: Added Claus, Lucas, Kumatora, Salsa, Erdrick, Link, Dark Link, Zelda, Sabin, Terra, Celes, Kefka, Ultros, Mog, Shadow, Interceptor, Mario, Luigi, Toad, Toadette, Bomberman, Captain Falcon, and Kirby.